Allow to set LRS endpoint adapter by setting `PupilfirstXapi.lrs` module variable.
If not set default one (`PupilfirstXapi::Lrs`) will be used.
The LRS adapter must implement a `call` method with single argument of XAPI
statement. The statement passed to LRS adapter could be `nil` and in this case
it must not send it to LRS store.
Useful in tests when you do not want any HTTP requests to be made. Just add
your own LRS adapter.